Finished Button Implementation.

+++ Anchored/UI/Constraints/CenterConstraint.cs
@@ -10,22 +10,22 @@ namespace Anchored.UI.Constraints
 		{
 		}
 
-		public override void ConstrainX(ref float value, UIComponent component)
+		public override void ConstrainX(ref int value, UIComponent component)
 		{
 			value = (Game1.WINDOW_WIDTH / 2) - (component.Width / 2);
 		}
 
-		public override void ConstrainY(ref float value, UIComponent component)
+		public override void ConstrainY(ref int value, UIComponent component)
 		{
 			value = (Game1.WINDOW_HEIGHT / 2) - (component.Height / 2);
 		}
 
-		public override void ConstrainWidth(ref float value, UIComponent component)
+		public override void ConstrainWidth(ref int value, UIComponent component)
 		{
 			throw new NotImplementedException();
 		}
 
-		public override void ConstrainHeight(ref float value, UIComponent component)
+		public override void ConstrainHeight(ref int value, UIComponent component)
 		{
 			throw new NotImplementedException();
 		}
