+++ Anchored/UI/UIConstraint.cs
@@ -6,9 +6,9 @@ namespace Anchored.UI
 {
 	public abstract class UIConstraint
 	{
-		public abstract void ConstrainX(ref float value, UIComponent component);
-		public abstract void ConstrainY(ref float value, UIComponent component);
-		public abstract void ConstrainWidth(ref float value, UIComponent component);
-		public abstract void ConstrainHeight(ref float value, UIComponent component);
+		public abstract void ConstrainX(ref int value, UIComponent component);
+		public abstract void ConstrainY(ref int value, UIComponent component);
+		public abstract void ConstrainWidth(ref int value, UIComponent component);
+		public abstract void ConstrainHeight(ref int value, UIComponent component);
 	}
 }
